requirements.txt
flask
clova-cek-sdk==1.1.0
